pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "orangebeard-robotframework"
version = "1.0.1"
authors = [
  { name="Tom Heintzberger", email="tom@orangebeard.io" },
]
description = "A robotframework listener for Orangebeard"
readme = "README.MD"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
        "License :: OSI Approved :: GNU Affero General Public License v3",
        "Operating System :: OS Independent",
]
dependencies = ["robotframework", "orangebeard-client"]

[project.urls]
"Homepage" = "https://github.com/orangebeard-io/robotframework-listener"
"Bug Tracker" = "https://github.com/orangebeard-io/robotframework-listener/issues"